requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning (for LSTM)
torch>=2.0.0
torchvision>=0.15.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Distributed Computing (for parallel backtesting)
ray>=2.7.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging
colorlog>=6.7.0
loguru>=0.7.0
